DatabaseIntelligence/services/query-agent/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
pydantic-settings==2.2.1
httpx==0.27.0
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
sqlparse==0.5.0

# DB driver extras (uncomment as needed for your target databases)
# pymysql==1.1.0          # MySQL
# pyodbc==5.1.0           # SQL Server
